Move API registry to package `core/discovery/api`
